Factored out BaseArbitraryWidget
